Distance-based outlier detection

+++ Python3Code/crowdsignals_ch3_outliers.py
@@ -130,16 +130,16 @@ if __name__ == '__main__':
                         'chauvenet' applies Chauvenet outlier detection method to a single variable \
                         'final' is used for the next chapter", choices=['LOF', 'distance', 'mixture', 'chauvenet', 'final'])
 
-    parser.add_argument('--C', type=int, default=2,
+    parser.add_argument('--C', type=float, default=2,
                         help="Chauvenet: C parameter")
    
     parser.add_argument('--K', type=int, default=5,
                         help="Local Outlier Factor:  K is the number of neighboring points considered")
 
-    parser.add_argument('--dmin', type=int, default=0.10,
+    parser.add_argument('--dmin', type=float, default=0.10,
                         help="Simple distance based:  dmin is ... ")
 
-    parser.add_argument('--fmin', type=int, default=0.99,
+    parser.add_argument('--fmin', type=float, default=0.99,
                         help="Simple distance based:  fmin is ... ")
 
     FLAGS, unparsed = parser.parse_known_args()
